Initial setup for the project
